refactor: :recycle: Move dataclasses into their own folder for better organization
