+++ py_now_playing/__init__.py
@@ -3,7 +3,7 @@ py_now_playing - A Python library for interacting with the Now Playing API.
 This library provides classes and functions to retrieve and manage information about the currently playing media,
 including playback status, media timeline, and media details.
 """
-from py_now_playing.media_timeline import MediaTimeline
-from py_now_playing.playback_info import PlaybackInfo, MediaPlaybackStatus, MediaPlaybackType, MediaPlaybackAutoRepeatMode
-from py_now_playing.pnp import PyNowPlaying
-from py_now_playing.media_info import MediaInfo
+from py_now_playing.dataclasses.media_timeline import MediaTimeline
+from py_now_playing.dataclasses.playback_info import PlaybackInfo, MediaPlaybackStatus, MediaPlaybackType, MediaPlaybackAutoRepeatMode
+from py_now_playing.core import PyNowPlaying
+from py_now_playing.dataclasses.media_info import MediaInfo
