+++ py_now_playing/__main__.py
@@ -1,6 +1,6 @@
 # from py_now_playing.media_timeline import MediaTimeline
 # from py_now_playing.playback_info import PlaybackInfo, MediaPlaybackStatus, MediaPlaybackType, MediaPlaybackAutoRepeatMode
-from py_now_playing.pnp import PyNowPlaying
+from py_now_playing.core import PyNowPlaying
 # from py_now_playing.media_info import MediaInfo
 import asyncio
 import argparse
