+++ py_now_playing/dataclasses/media_info.py
